# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create a Stripe checkout session
      security:
        - cookie: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                priceId:
                  type: string
                  description: Stripe price ID. Falls back to hard-coded $12/mo when omitted.
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required: [url]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /billing/confirm:
    get:
      operationId: confirmCheckout
      tags: [billing]
      summary: Server-verify a completed Stripe checkout session and upgrade tier
      security:
        - cookie: []
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tier upgraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Payment not completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /billing/portal:
    post:
      operationId: createBillingPortal
      tags: [billing]
      summary: Create a Stripe billing portal session
      security:
        - cookie: []
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required: [url]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: sovv_session
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...